Instagram Fake Account Detection Project
The aim of this project is to build and train a deep neural network model to detect fake or spam instagram accounts

This model is trained such that it considers the above given features and determines whether a particular account is fake or not. By resulting the output as either 0 or 1 meaning TRUSTED or FAKE respectively. Our intention is to make this software capable of thinking like a human, based on the data it is given and results in maximum probability of success.